requirements.txt
Django==2.2.11
django-cors-headers==3.2.1
djangorestframework==3.11.0
mysqlclient==1.4.6
Pillow==7.0.0
python-decouple==3.3
pytz==2019.3
sqlparse==0.3.1
gunicorn==20.0.4
psycopg2-binary==2.8.4
